Re-enable code coverage upload
