+++ tests/framework/e2e/etcd_spawn_nocov.go
@@ -18,7 +18,6 @@
 package e2e
 
 import (
-	"fmt"
 	"os"
 	"strings"
 
@@ -27,10 +26,6 @@ import (
 )
 
 const noOutputLineCount = 0 // regular binaries emit no extra lines
-
-func SpawnCmd(args []string, envVars map[string]string) (*expect.ExpectProcess, error) {
-	return SpawnCmdWithLogger(zap.NewNop(), args, envVars)
-}
 
 func SpawnCmdWithLogger(lg *zap.Logger, args []string, envVars map[string]string) (*expect.ExpectProcess, error) {
 	wd, err := os.Getwd()
@@ -45,24 +40,4 @@ func SpawnCmdWithLogger(lg *zap.Logger, args []string, envVars map[string]string
 	}
 	lg.Info("spawning process", zap.Strings("args", args), zap.String("working-dir", wd), zap.Strings("environment-variables", env))
 	return expect.NewExpectWithEnv(args[0], args[1:], env)
-}
-
-func mergeEnvVariables(envVars map[string]string) []string {
-	var env []string
-	// Environment variables are passed as parameter have higher priority
-	// than os environment variables.
-	for k, v := range envVars {
-		env = append(env, fmt.Sprintf("%s=%s", k, v))
-	}
-
-	// Now, we can set os environment variables not passed as parameter.
-	currVars := os.Environ()
-	for _, v := range currVars {
-		p := strings.Split(v, "=")
-		if _, ok := envVars[p[0]]; !ok {
-			env = append(env, fmt.Sprintf("%s=%s", p[0], p[1]))
-		}
-	}
-
-	return env
 }
